feat: split articles and tweets into separate pages
